requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
yt-dlp==2023.12.30
lxml==4.9.3 